running status bugfix
Some MIDI interfaces send a "running status" (omitting the first message
byte, if it is the same for consecutive MIDI messages, even if the
application doesn't send it. The current menu doesn't support this, so
as a workaround it alternates bewteen note-on/off messages, because
after the first note-off message it is not tested if it is a note-on for
the rest of the transfer.